created individual project component
