+++ web/src/router/index.js
@@ -3,6 +3,7 @@ import VueRouter from 'vue-router';
 import AuthAPI from '@/utils/api/auth';
 import AboutPage from '@/components/AboutPage';
 import ResumePage from '@/components/ResumePage';
+import ProjectPage from '@/components/ProjectPage';
 import ProjectsPage from '@/components/ProjectsPage';
 import BuildsPage from '@/components/BuildsPage';
 import LoginPage from '@/components/LoginPage';
@@ -16,7 +17,6 @@ import { API_TOKEN_KEY } from '@/store/constants';
 
 Vue.use(VueRouter);
 
-let initialPageLoad = null;
 const defaultTitle = 'Chris Meyers - Developer, Tech Enthusiast';
 
 const formatPageTitle = (name) => {
@@ -40,17 +40,7 @@ const formatDashboardTitle = (name) => {
 const router = new VueRouter({
   mode: 'history',
   scrollBehavior(to, from, savedPosition) {
-    if (to.hash) {
-      return new Promise((resolve) => {
-        const timeout = initialPageLoad ? 400 : 100;
-        setTimeout(() => {
-          document
-            .getElementById(to.hash.replace('#', ''))
-            .scrollIntoView(true);
-          resolve(true);
-        }, timeout);
-      });
-    } else if (savedPosition) {
+    if (savedPosition) {
       return savedPosition;
     } else {
       return { x: 0, y: 0 };
@@ -74,16 +64,24 @@ const router = new VueRouter({
       },
     },
     {
-      path: '/builds',
-      component: BuildsPage,
+      path: '/projects',
+      component: ProjectsPage,
       meta: {
         secure: false,
         title: defaultTitle,
       },
     },
     {
-      path: '/projects',
-      component: ProjectsPage,
+      path: '/projects/:id',
+      component: ProjectPage,
+      meta: {
+        secure: false,
+        title: defaultTitle,
+      },
+    },
+    {
+      path: '/builds',
+      component: BuildsPage,
       meta: {
         secure: false,
         title: defaultTitle,
@@ -151,12 +149,6 @@ const router = new VueRouter({
 });
 
 router.beforeEach(async (to, from, next) => {
-  if (initialPageLoad === null) {
-    initialPageLoad = true;
-  } else if (initialPageLoad === true) {
-    initialPageLoad = false;
-  }
-
   let match, authorized;
 
   // Update the HTML title
