+++ web/src/utils/api/projects.js
@@ -19,6 +19,21 @@ export default {
       });
   },
 
+  getById(id, params = {}) {
+    const queryString = qs.stringify(params, {
+      addQueryPrefix: true,
+      strictNullHandling: true,
+    });
+    return axios
+      .get(`/public/projects/${id}${queryString}`)
+      .then((response) => {
+        return response;
+      })
+      .catch((error) => {
+        return ErrorHandler.handle(error);
+      });
+  },
+
   // POST Methods
   add(token, project) {
     return axios({
